Implement semantic analysis for the compiler
- Added SemanticAnalyzer class to perform semantic checks on the AST.
- Implemented error and warning reporting mechanisms.
- Developed scope management to handle variable and function declarations.
- Created visitors for various AST nodes including declarations, expressions, and statements.
- Introduced type utility functions to check type compatibility and qualifications.
- Enhanced handling of enums, structs, and typedefs in the semantic analysis phase.
- Implemented expression type inference for literals, identifiers, and binary/unary operations.
- Added support for function declarations and parameter handling in semantic analysis.
