+++ include/semantics/SemanticAnalyzer.hpp
@@ -6,7 +6,6 @@
 #include <utils/SourceLocation.hpp>
 #include <string>
 #include <map>
-#include <set>
 #include <vector>
 #include <memory>
